Module mongo_driver::collection

Access to a MongoDB collection.

Collection is the main type used when accessing collections.

Structs

AggregateOptions
Options to configure an aggregate operation.

BulkOperation
Provides an abstraction for submitting multiple write operations as a single batch.

BulkOperationOptions
Options to configure a bulk operation.

Collection
Provides access to a collection for most CRUD operations, I.e. insert, update, delete, find, etc.

CountOptions
Options to configure a count operation.

FindAndModifyOptions
Options to configure a find and modify operation.

InsertOptions
Options to configure an insert operation.

RemoveOptions
Options to configure a remove operation.

TailOptions
Options to configure a tailing query.

UpdateOptions
Options to configure an update operation.

Enums

FindAndModifyOperation
Possible find and modify operations.