Struct mongo_driver::CommandAndFindOptions

pub struct CommandAndFindOptions {
    pub query_flags: Flags<QueryFlag>,
    pub skip: u32,
    pub limit: u32,
    pub batch_size: u32,
    pub fields: Option<Document>,
    pub read_prefs: Option<ReadPrefs>,
}

Options to configure both command and find operations.

Fields

query_flags

Flags to use

skip

Number of documents to skip, zero to ignore

limit

Max number of documents to return, zero to ignore

batch_size

Number of documents in each batch, zero to ignore (default is 100)

fields

Fields to return, not all commands support this option

read_prefs

Read prefs to use

Methods

impl CommandAndFindOptions

fn default() -> CommandAndFindOptions

Default options used if none are provided.

fn with_fields(fields: Document) -> CommandAndFindOptions