Struct mongo_driver::client::Uri

pub struct Uri {
    // some fields omitted
}

Abstraction on top of MongoDB connection URI format.

Methods

impl Uri

fn new<T: Into<Vec<u8>>>(uri_string: T) -> Option<Uri>

Parses a string containing a MongoDB style URI connection string.

Returns None if the uri is not in the correct format, there is no further information available if this is not the case.

fn as_str<'a>(&'a self) -> Cow<'a, str>

fn get_database<'a>(&'a self) -> Option<Cow<'a, str>>